pyproject.toml
[tool.poetry]
name = "decodeb64"
version = "0.1.0"
description = ""
authors = ["Alisson R. Perez <756802+alissonperez@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
PyYAML = "^6.0.2"


[tool.poetry.group.dev.dependencies]
icecream = "^2.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
